% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escapeLUR.R
\name{escapeLUR}
\alias{escapeLUR}
\title{Function for deriving parametric land use regression (LUR) model according
   to the ESCAPE procedure.}
\usage{
escapeLUR(data, pred, depVar, dirEff, thresh = 0.95)
}
\arguments{
\item{data}{A data set which contains the dependent variable and the
potential predictors.}

\item{pred}{A character vector stating the variable names of the
potential predictors (names have to match the column names of
`data`).}

\item{depVar}{A character string indicating the name of the dependent
variable.}

\item{dirEff}{A vector that contains one entry for each potential
predictor and indicates the expected direction of the effect of the
potential predictor (1 for positive, -1 for negative and 0 if the
expected effect sign is unclear).}

\item{thresh}{A numeric value that indicates the maximum share of
zero values; if the share is exceeded, the corresponding potential
predictor is excluded.}
}
\value{
An object of class `escapeLUR` with the following elements:

\item{coefficients}{a vector containing the coefficient estimates}

It has `...`, `...`, and `...` methods.
}
\description{
\code{escapeLUR} fits a parametric land use regression (LUR) model according
   to the ESCAPE procedure. The procedure automatically carries out the
   procedure outlined in
   \insertCite{Beelen.2013;textual}{smoothLUR},
   \insertCite{Eeftens.2016;textual}{smoothLUR}, and
   \insertCite{Wolf.2017;textual}{smoothLUR}.
}
\examples{
## Load data from package
data(monSitesDE, package="smoothLUR")
set.seed(42)

## Code example
dat <- monSitesDE[sample(1:nrow(monSitesDE), 40),]
m1 <- escapeLUR(data = dat
                 ,pred = c("AQeLon", "AQeLat", "AQeAlt", "HighDens"
                         ,"LowDens", "Ind", "Transp", "Seap", "Airp", "Constr"
                         ,"UrbGreen", "Agri", "Forest", "BBSRpopDens"
                         ,"PriRoad", "SecRoad", "NatMot", "LocRoute")
                 ,depVar = "AQeYMean"
                 ,dirEff = c(0,0,-1,1,1,1,1,1,1,1,-1,0,-1,1,1,1,1,1)
                 ,thresh = 0.95)

summary(m1)
summary(m1)$adj.r.squared
BIC(m1)
AIC(m1)

\donttest{
## Load data from package
data(monSitesDE, package="smoothLUR")
dat <- monSitesDE
m1 <- escapeLUR(data = dat
                 ,pred = c("AQeLon", "AQeLat", "AQeAlt", "HighDens"
                         ,"LowDens", "Ind", "Transp", "Seap", "Airp", "Constr"
                         ,"UrbGreen", "Agri", "Forest", "BBSRpopDens"
                         ,"PriRoad", "SecRoad", "NatMot", "LocRoute")
                 ,depVar = "AQeYMean"
                 ,dirEff = c(0,0,-1,1,1,1,1,1,1,1,-1,0,-1,1,1,1,1,1)
                 ,thresh = 0.95)

summary(m1)
summary(m1)$adj.r.squared
BIC(m1)
AIC(m1) 

}
 
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{smoothLUR}} for smooth land use regression (LUR)
   models.
\code{\link{kFoldCV}} for k-fold cross-validation for
   escapeLUR and smoothLUR objects.
\code{\link{looCV}} for leave-one-out cross-validation for
   escapeLUR and smoothLUR objects.
}
\author{
Svenia Behm and Markus Fritsch
}
