% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parLUR.R
\name{parLUR}
\alias{parLUR}
\title{Function for deriving parametric land use regression (LUR) model.}
\usage{
parLUR(data, x, y, dirEff, thresh = 0.95, thresh_pval = 0.1)
}
\arguments{
\item{data}{A data set which contains the dependent variable and the
potential predictors.}

\item{x}{A character vector stating the variable names of the
potential predictors (names have to match the column names of
`data`).}

\item{y}{A character string indicating the name of the dependent
variable (name needs to match the column names of `data`).}

\item{dirEff}{A vector that contains one entry for each potential
predictor and indicates the expected direction of the effect of the
potential predictor (1 for positive, -1 for negative and 0 if the
expected effect sign is unclear).}

\item{thresh}{A numeric value that indicates the maximum share of
zero values; if the share is exceeded, the corresponding potential
predictor is excluded (defaults to 0.95).}

\item{thresh_pval}{A numeric value that indicates a threshold for
removing predictors after the forward stepwise modeling procedure;
the threshold is the p-value of a standard t-Test (defaults to 0.1).}
}
\value{
An object of class `parLUR` with the following elements:

\item{coefficients}{a vector containing the coefficient estimates}

It has `...`, `...`, and `...` methods.
}
\description{
\code{parLUR} fits a parametric land use regression (LUR) model employing
   parametric polynomials in a forward stepwise approach. The function
   allows to carry out the ESCAPE procedure outlined in
   \insertCite{Beelen.2013;textual}{smoothLUR},
   \insertCite{Eeftens.2016;textual}{smoothLUR},
   \insertCite{Wolf.2017;textual}{smoothLUR}, and
   \insertCite{Fritsch2020smooth;textual}{smoothLUR}.
}
\examples{
## Load data from package
data(monSitesDE, package="smoothLUR")
set.seed(42)

## Code example
dat <- monSitesDE[sample(1:nrow(monSitesDE), 40),]
m1 <- parLUR(data = dat
                 ,x = c("Lon", "Lat", "Alt", "HighDens"
                         ,"LowDens", "Ind", "Transp", "Seap", "Airp", "Constr"
                         ,"UrbGreen", "Agri", "Forest", "PopDens"
                         ,"PriRoad", "SecRoad", "FedAuto", "LocRoute")
                 ,y = "Y"
                 ,dirEff = c(0,0,-1,1,1,1,1,1,1,1,-1,0,-1,1,1,1,1,1)
                 ,thresh = 0.95
                 ,thresh_pval = 0.1)

summary(m1)
summary(m1)$adj.r.squared
BIC(m1)
AIC(m1)

\donttest{
## Load data from package
data(monSitesDE, package="smoothLUR")
dat <- monSitesDE
m1 <- parLUR(data = dat
                 ,x = c("Lon", "Lat", "Alt", "HighDens"
                         ,"LowDens", "Ind", "Transp", "Seap", "Airp", "Constr"
                         ,"UrbGreen", "Agri", "Forest", "PopDens"
                         ,"PriRoad", "SecRoad", "FedAuto", "LocRoute")
                 ,y = "Y"
                 ,dirEff = c(0,0,-1,1,1,1,1,1,1,1,-1,0,-1,1,1,1,1,1)
                 ,thresh = 0.95
                 ,thresh_pval = 0.1)

summary(m1)
summary(m1)$adj.r.squared
BIC(m1)
AIC(m1)

}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{smoothLUR}} for smooth land use regression (LUR)
   models.
\code{\link{kFoldCV}} for k-fold cross-validation for
   parLUR and smoothLUR objects.
}
\author{
Svenia Behm and Markus Fritsch
}
