% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-smoothLUR.R, R/smoothLUR.R
\docType{package}
\name{smoothLUR}
\alias{smoothLUR}
\title{smoothLUR: A package for smooth land use regression modeling}
\usage{
smoothLUR(data, pred, spVar1, spVar2, depVar, thresh = 0.95)
}
\arguments{
\item{data}{A data set which contains the dependent variable and
the potential predictors.}

\item{pred}{A character vector stating the variable names of the
potential predictors (names have to match the column names of
`data`).}

\item{spVar1}{A character vector stating the variable name referring
to longitude (name has to match the column name of `data`)}

\item{spVar2}{A character vector stating the variable name referring
to latitude (name has to match the column name of `data`)}

\item{depVar}{A character string indicating the name of the dependent
variable.}

\item{thresh}{A numeric value that indicates the maximum share of
zero values; if the share is exceeded, the corresponding potential
predictor is excluded.}
}
\value{
An object of class `smoothLUR` with the following elements:

\item{coefficients}{a vector containing the coefficient estimates}

It has `...`, `...`, and `...` methods.
}
\description{
The smoothLUR package provides function for smooth land use
regression modeling and parametric benchmarks. The main
functions of the package are \code{smoothLUR}, \code{escapeLUR},
and the evaluation functions \code{loocvLUR} and \code{kFoldcvLUR}.

\code{smoothLUR} fits a smooth land use regression (LUR) model
    using the gam() function from the mgcv package. The procedure
    automatically carries out the procedure outlined in ...
}
\examples{
\dontrun{
## Load data set
data(monSitesDE, package="smoothLUR")
set.seed(42)

## Code example
dat <- monSitesDE[sample(1:nrow(monSitesDE), 40),]
m1 <- smoothLUR(data = dat
                 ,pred = c("AQeLon", "AQeLat", "AQeAlt", "HighDens"
                         ,"LowDens", "Ind", "Transp", "Seap", "Airp", "Constr"
                         ,"UrbGreen", "Agri", "Forest", "BBSRpopDens"
                         ,"PriRoad", "SecRoad", "NatMot", "LocRoute")
                 ,spVar1 = "AQeLon"
                 ,spVar2 = "AQeLat"
                 ,depVar = "AQeYMean"
                 ,thresh = 0.95)

summary(m1)
summary(m1)$adj.r.squared
BIC(m1)
AIC(m1)

\donttest{
## Load data set
data(monSitesDE, package="smoothLUR")
dat <- monSitesDE
m1 <- smoothLUR(data = dat,
                 ,pred = c("AQeLon", "AQeLat", "AQeAlt", "HighDens"
                         ,"LowDens", "Ind", "Transp", "Seap", "Airp", "Constr"
                         ,"UrbGreen", "Agri", "Forest", "BBSRpopDens"
                         ,"PriRoad", "SecRoad", "NatMot", "LocRoute")
                 ,spVar1 = "AQeLon"
                 ,spVar2 = "AQeLat"
                 ,depVar = "AQeYMean"
                 ,thresh = 0.95)

summary(m1)
summary(m1)$adj.r.squared
BIC(m1)
AIC(m1) 

}
}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{escapeLUR}} for smooth land use regression (LUR)
   models.
\code{\link{kFoldCV}} for k-fold cross-validation for
   escapeLUR and smoothLUR objects.
\code{\link{looCV}} for leave-one-out cross-validation for
   escapeLUR and smoothLUR objects.
}
\author{
Svenia Behm and Markus Fritsch
}
