% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfcv.R
\name{kFoldCV}
\alias{kFoldCV}
\title{Function for conducting k-fold cross-validation}
\usage{
kFoldCV(
  data,
  pred,
  ID,
  spVar1,
  spVar2,
  depVar,
  dirEff,
  thresh = 0.95,
  seed,
  k = 10,
  strat = FALSE
)
}
\arguments{
\item{data}{A data set which contains the dependent variable and
the potential predictors.}

\item{pred}{A character vector stating the variable names of the
potential predictors (names have to match the column names of
`data`).}

\item{ID}{A character string stating the variable name referring
to the monitoring sites' ID (name has to mach the column name
of `data`)}

\item{spVar1}{A character string stating the variable name
referring to longitude (name has to match the column name of
`data`)}

\item{spVar2}{A character string stating the variable name
referring to latitude (name has to match the column name of
`data`)}

\item{depVar}{A character string that indicates the name of the
dependent variable.}

\item{dirEff}{A vector that contains one entry for each potential
predictor and indicates the expected direction of the effect
of the potential predictor (1 for positive, -1 for negative
and 0 if the expected effect sign is unclear). Argument
defaults to NULL and is only required for parametric model
fitting according to ESCAPE procedure.}

\item{thresh}{A numeric value that indicates the maximum share of
zero values; if the share is exceeded, the corresponding
potential predictor is excluded.}

\item{seed}{A numeric value that defines the seed for random
sampling}

\item{strat}{A boolean value that indicates whether stratified
sampling is desired (stratified spatially w.r.t. German federal
states)}

\item{folds}{An integer denoting the number of folds to use in
cross-validation (defaults to 10)}
}
\value{
An object of class `tfcvLUR` with the following elements:

\item{df.err}{data.frame with four columns: ID (Id of monitoring
   site), Fold (Number of fold the monitoring site is attributed to),
   Err.par (Errors derived from parametric LUR model), Err.smooth
   (Errors derived from smooth LUR model)}
\item{ls.models}{list with elements according to `folds'; each list
   element is named according to the omitted fold and is itself a
   list containing two elements: mod.par (parametric model based
   on remaining sites), mod.smooth (smooth model based on remaining
   sites)}

It has `...`, `...`, and `...` methods.
}
\description{
\code{kFoldCV} conducts a k-fold cross-validation for parametric
   and smooth land use regression (LUR) models fitted with the
   functions \code{escapeLUR} and \code{smoothLUR}, respectively.
}
\examples{
## Load data set
dat <- data("monSitesDE", package="smoothLUR")

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{escapeLUR}} for parametric land use regression (LUR)
   modeling.
\code{\link{smoothLUR}} for smooth land use regression (LUR)
   modeling.
\code{\link{looCV}} for leave-one-out cross-validation for
   `escapeLUR` and `smoothLUR` objects.
}
\author{
Svenia Behm and Markus Fritsch
}
