% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-smoothLUR.R, R/smoothLUR.R
\docType{package}
\name{smoothLUR}
\alias{smoothLUR}
\title{smoothLUR: A package for smooth land use regression modeling}
\usage{
smoothLUR(data, x, spVar1, spVar2, y, thresh = 0.95)
}
\arguments{
\item{data}{A data set which contains the dependent variable and
the potential predictors.}

\item{x}{A character vector stating the variable names of the
potential predictors (names have to match the column names of
`data`).}

\item{spVar1}{A character vector stating the variable name referring
to longitude (name has to match the column name of `data`).}

\item{spVar2}{A character vector stating the variable name referring
to latitude (name has to match the column name of `data`).}

\item{y}{A character string indicating the name of the dependent
variable (name has to match the column name of `data`).}

\item{thresh}{A numeric value that indicates the maximum share of
zero values; if the share is exceeded, the corresponding potential
predictor is excluded.}
}
\value{
An object of class `smoothLUR` with the following elements:

\item{coefficients}{a vector containing the coefficient estimates}

It has `...`, `...`, and `...` methods.
}
\description{
The smoothLUR package provides functions for smooth land use
regression modeling and parametric benchmarks to reproduce
the results reported in Fritsch and Behm (2020).
The main functions of the package are \code{smoothLUR},
\code{parLUR}, and the evaluation functions \code{kFoldCV} and
\code{looCV}.

\code{smoothLUR} fits a smooth land use regression (LUR) model
    using the gam() function from the mgcv package. The procedure
    is outlined in \insertCite{Fritsch2020smooth;textual}{smoothLUR}
}
\examples{
\dontrun{
## Load data set
data(monSitesDE, package="smoothLUR")
set.seed(42)

## Code example
dat <- monSitesDE[sample(1:nrow(monSitesDE), 40),]
m1 <- smoothLUR(data = dat
                 ,x = c("Lon", "Lat", "Alt", "HighDens"
                         ,"LowDens", "Ind", "Transp", "Seap", "Airp", "Constr"
                         ,"UrbGreen", "Agri", "Forest", "PopDens"
                         ,"PriRoad", "SecRoad", "FedAuto", "LocRoute")
                 ,spVar1 = "Lon"
                 ,spVar2 = "Lat"
                 ,y = "Y"
                 ,thresh = 0.95)

summary(m1)
summary(m1)$adj.r.squared
BIC(m1)
AIC(m1)

\donttest{
## Load data set
data(monSitesDE, package="smoothLUR")
dat <- monSitesDE
m1 <- smoothLUR(data = dat,
                 ,x = c("Lon", "Lat", "Alt", "HighDens"
                         ,"LowDens", "Ind", "Transp", "Seap", "Airp", "Constr"
                         ,"UrbGreen", "Agri", "Forest", "PopDens"
                         ,"PriRoad", "SecRoad", "FedAuto", "LocRoute")
                 ,spVar1 = "Lon"
                 ,spVar2 = "Lat"
                 ,y = "Y"
                 ,thresh = 0.95)

summary(m1)
summary(m1)$adj.r.squared
BIC(m1)
AIC(m1)

}
}
}
\references{
Fritsch and Behm (2020): Modeling the effects of agglomeration
areas and connecting infrastructure in land use regression --
Specification, estimation, and interpretations. Working paper.

\insertAllCited{}
}
\seealso{
\code{\link{parLUR}} for parametric land use regression (LUR)
   models.
\code{\link{kFoldCV}} for k-fold cross-validation for
   parLUR and smoothLUR objects.
}
\author{
Svenia Behm and Markus Fritsch
}
